Extract steps to custom components
