+++ client/components/jetpack/jetpack-welcome-page/index.tsx
@@ -5,18 +5,12 @@ import Main from 'calypso/components/main';
 
 import './style.scss';
 
-export type WelcomeStep = {
-	content?: React.ReactNode;
-	id?: string;
-	title?: React.ReactNode;
-};
-
 export type JetpackWelcomePageProps = {
 	description?: React.ReactNode;
 	footer?: React.ReactNode;
 	mainClassName?: string;
 	pageViewTracker?: React.ReactNode;
-	steps?: Array< WelcomeStep >;
+	steps?: Array< React.ReactNode >;
 	title: React.ReactNode;
 };
 
@@ -37,16 +31,11 @@ export const JetpackWelcomePage: React.FC< JetpackWelcomePageProps > = ( {
 					<h1 className="jetpack-welcome-page__title">{ title }</h1>
 					<p className="jetpack-welcome-page__description">{ description }</p>
 					<div className="jetpack-welcome-page__steps">
-						{ steps?.map( ( { content, id, title }, index ) => {
+						{ steps?.map( ( step, index ) => {
 							return (
-								<div className="jetpack-welcome-page__step" key={ id || index }>
+								<div className="jetpack-welcome-page__step" key={ index }>
 									<div className="jetpack-welcome-page__step--number">{ index + 1 }</div>
-									<div className="jetpack-welcome-page__step--body">
-										{ title ? (
-											<h2 className="jetpack-welcome-page__step--title">{ title }</h2>
-										) : null }
-										<div className="jetpack-welcome-page__step--content">{ content }</div>
-									</div>
+									<div className="jetpack-welcome-page__step--content">{ step }</div>
 								</div>
 							);
 						} ) }
