+++ client/components/jetpack/jetpack-welcome-page/style.scss
@@ -66,11 +66,6 @@
 				padding-bottom: 0;
 			}
 
-			&--body {
-				flex: 1;
-				padding: 0 10px 0 16px;
-			}
-
 			&--number {
 				color: white;
 				background-color: var( --studio-jetpack-green-40 );
@@ -83,6 +78,7 @@
 			}
 			&--content {
 				flex: 1;
+				padding: 0 10px 0 16px;
 				p {
 					@include break-mobile {
 						font-size: 18px; /* stylelint-disable-line */
